# taskboard/public/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TaskBoard API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing Kanban tasks with integrated MCP (Model Context Protocol) chat functionality.\n\n## Authentication\n\nProtected endpoints require user identification via one of these methods:\n1. **Bearer Token**: `Authorization: Bearer <userId>`\n2. **Query Parameter**: `?userId=<userId>`  \n3. **Request Body**: Include `userId` field in JSON body\n\nThis flexible authentication supports both web applications and MCP integrations.\n\n## Features\n\n- User registration and authentication\n- CRUD operations for Kanban items (todo, doing, done)  \n- MCP-powered AI chat assistant integration\n- Session-based conversation management\n- Tool usage tracking and display"
  },
  "servers": [
    {
      "url": "${API_BASE_URL:-http://localhost:3000}",
      "description": "API server (configurable via API_BASE_URL environment variable)"
    }
  ],
...
  "paths": {
...
    "/api/login": {
      "post": {
        "summary": "Authenticate user",
        "description": "Login with username and password to receive user ID token",
        "operationId": "user_login",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "x-gram": {
          "name": "login_user",
          "description": "<context>\nThis endpoint authenticates a user with their username and password.\nUpon successful authentication, it returns a user ID that serves as the authentication token.\n</context>\n<prerequisites>\n- User must have an existing account (use register endpoint to create one)\n- Username and password are required\n</prerequisites>\n<usage>\n- Use the returned userId as a query parameter for subsequent API calls\n- Store the userId securely as it provides access to the user's data\n- Pass the userId as a query parameter in all authenticated requests\n</usage>\n",
          "responseFilterType": "jq"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCredentials"
              },
              "example": {
                "username": "john_doe",
                "password": "secure_password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "userId": "clh7x1234abcd5678efgh"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing username or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Missing username or password"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Invalid credentials"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Create a new user account with username and password",
        "operationId": "user_register",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "x-gram": {
          "name": "register_user",
          "description": "<context>\nThis endpoint creates a new user account with a username and password.\nUpon successful registration, it returns a user ID that serves as the authentication token.\n</context>\n<prerequisites>\n- Username must be unique (3-50 characters)\n- Password must be secure (6-100 characters)\n- Username and password are required\n</prerequisites>\n<usage>\n- Use this to create new user accounts\n- The returned userId serves as the authentication token\n- Store the userId securely as it provides access to the user's data\n- Pass the userId as a query parameter in all authenticated requests\n</usage>\n",
          "responseFilterType": "jq"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCredentials"
              },
              "example": {
                "username": "john_doe",
                "password": "secure_password123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "userId": "clh7x1234abcd5678efgh"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing username or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Missing username or password"
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "User already exists"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserCredentials": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "description": "Unique username for the account",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 100,
            "description": "Password for the account",
            "example": "secure_password123"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": [
          "userId"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "description": "Unique identifier for the user (used as bearer token)",
            "example": "clh7x1234abcd5678efgh"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    },
    "securitySchemes": {
      "UserIdParam": {
        "type": "apiKey",
        "in": "query",
        "name": "userId",
        "description": "User ID parameter for authentication"
      }
    }
  }
}